PFA = Projet de Fin d'année
It is a Projet that us IT engineering students have to do in our second year, it is mostly a warmup for our final graduation projet which is called " PFE ".

THis project is about using symfony to develop a Website for managing the ads for used cars. It it a whole platform made with web development tools.

The project is based on symfony 3.2 and doesn't integrate any client side work. it uses TWIG as a templating engine

Using this project in your personal matters is not forbidden but i would appreciate the repository to be credited. Thank you.